+++ docs/ChangeLog.md
@@ -189,6 +189,7 @@
 - decode(`ble/builtin/bind`): support single quotes in the macro/command strings `#D2236` 2f90120e
 - mandb: process less formatting sequences in parsing `--help` `#D2244` 60d36ba5
 - mandb: hook into bash-completion's `_comp_command_offset` `#D2255` cbcce625
+- canvas: update tables for Unicode 16.0.0 `#D2283` xxxxxxxx xxxxxxxx
 
 ## Contrib